requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary==2.9.6
celery==5.3.0
redis==4.5.1
django-cors-headers==3.13.0
python-dotenv==0.21.1
django-debug-toolbar==4.4.6
